back/requirements.txt
Flask==2.1.2
flask-cors==3.0.10
neo4j==4.4.2
python-dotenv==0.21.0
